Split cesium classes into multiple files.
