+++ entwine/formats/cesium/tileset.cpp
@@ -8,9 +8,9 @@
 *
 ******************************************************************************/
 
+#include <entwine/formats/cesium/pnts.hpp>
+#include <entwine/formats/cesium/tile.hpp>
 #include <entwine/formats/cesium/tileset.hpp>
-#include <entwine/io/io.hpp>
-#include <entwine/types/binary-point-table.hpp>
 
 namespace entwine
 {
@@ -157,173 +157,6 @@ Json::Value Tileset::build(
     }
 
     return json;
-}
-
-Tile::Tile(const Tileset& tileset, const ChunkKey& ck, bool external)
-    : m_tileset(tileset)
-{
-    m_json["boundingVolume"]["box"] = toBox(ck.bounds());
-    m_json["geometricError"] = m_tileset.geometricErrorAt(ck.depth());
-    m_json["content"]["url"] = external ?
-        "tileset-" + ck.toString() + ".json" :
-        ck.toString() + ".pnts";
-    if (!ck.depth()) m_json["refine"] = "ADD";
-}
-
-Pnts::Pnts(const Tileset& tileset, const ChunkKey& ck)
-    : m_tileset(tileset)
-    , m_key(ck)
-{
-    m_mid = m_key.bounds().mid();
-
-    if (const Delta* d = m_tileset.metadata().delta())
-    {
-        m_mid = Point::unscale(m_mid, d->scale(), d->offset());
-    }
-}
-
-std::vector<char> Pnts::build()
-{
-    const auto data = m_tileset.metadata().dataIo().read(
-            m_tileset.in(),
-            m_tileset.tmp(),
-            m_tileset.pointPool(),
-            m_key.get().toString());
-
-    m_np = data.size();
-
-    const auto xyz(buildXyz(data));
-    const auto rgb(buildRgb(data));
-    return build(xyz, rgb);
-}
-
-Pnts::Xyz Pnts::buildXyz(const Cell::PooledStack& cells) const
-{
-    Xyz xyz;
-    xyz.reserve(m_np * 3);
-
-    Scale scale(1);
-    Offset offset(0);
-
-    if (const Delta* d = m_tileset.metadata().delta())
-    {
-        scale = d->scale();
-        offset = d->offset();
-    }
-
-    for (const auto& cell : cells)
-    {
-        const Point p(Point::unscale(cell.point(), scale, offset));
-        xyz.push_back(p.x - m_mid.x);
-        xyz.push_back(p.y - m_mid.y);
-        xyz.push_back(p.z - m_mid.z);
-    }
-
-    return xyz;
-}
-
-Pnts::Rgb Pnts::buildRgb(const Cell::PooledStack& cells) const
-{
-    Rgb rgb;
-    rgb.reserve(m_np * 3);
-
-    using DimId = pdal::Dimension::Id;
-    BinaryPointTable table(m_tileset.metadata().schema());
-
-    auto getByte([this, &table](DimId id) -> uint8_t
-    {
-        if (!m_tileset.truncate()) return table.ref().getFieldAs<uint8_t>(id);
-        else return table.ref().getFieldAs<uint16_t>(id) >> 8;
-    });
-
-    uint8_t r, g, b;
-
-    if (m_tileset.colorType() == ColorType::Tile)
-    {
-        r = std::rand() % 256;
-        g = std::rand() % 256;
-        b = std::rand() % 256;
-    }
-
-    assert(m_tileset.colorType() != ColorType::None);
-    for (const auto& cell : cells)
-    {
-        table.setPoint(cell.uniqueData());
-        if (m_tileset.colorType() == ColorType::Rgb)
-        {
-            r = getByte(DimId::Red);
-            g = getByte(DimId::Green);
-            b = getByte(DimId::Blue);
-        }
-        else if (m_tileset.colorType() == ColorType::Intensity)
-        {
-            r = g = b = getByte(DimId::Intensity);
-        }
-
-        rgb.push_back(r);
-        rgb.push_back(g);
-        rgb.push_back(b);
-    }
-
-    return rgb;
-}
-
-std::vector<char> Pnts::build(const Xyz& xyz, const Rgb& rgb) const
-{
-    Json::Value featureTable;
-    featureTable["POINTS_LENGTH"] = static_cast<Json::UInt64>(m_np);
-    featureTable["POSITION"]["byteOffset"] = 0;
-    featureTable["RTC_CENTER"] = m_mid.toJson();
-
-    if (m_tileset.hasColor())
-    {
-        featureTable["RGB"]["byteOffset"] = static_cast<Json::UInt64>(
-                xyz.size() * sizeof(float));
-    }
-
-    std::string featureString = toFastString(featureTable);
-    while (featureString.size() % 8) featureString += ' ';
-
-    const uint64_t headerSize(28);
-    const uint64_t binaryBytes = xyz.size() * (sizeof(float)) + rgb.size();
-    const uint64_t totalBytes = headerSize + featureString.size() + binaryBytes;
-
-    std::vector<char> header;
-    header.reserve(headerSize);
-
-    auto push([&header](uint32_t v)
-    {
-        header.insert(
-                header.end(),
-                reinterpret_cast<char*>(&v),
-                reinterpret_cast<char*>(&v + 1));
-    });
-
-    const std::string magic("pnts");
-    header.insert(header.end(), magic.begin(), magic.end());
-    push(1);                    // Version.
-    push(totalBytes);           // ByteLength.
-    push(featureString.size()); // FeatureTableJsonByteLength.
-    push(binaryBytes);          // FeatureTableBinaryByteLength.
-    push(0);                    // BatchTableJsonByteLength.
-    push(0);                    // BatchTableBinaryByteLength.
-    assert(header.size() == headerSize);
-
-    std::vector<char> pnts;
-    pnts.reserve(totalBytes);
-
-    pnts.insert(pnts.end(), header.begin(), header.end());
-    pnts.insert(pnts.end(), featureString.begin(), featureString.end());
-    pnts.insert(
-            pnts.end(),
-            reinterpret_cast<const char*>(xyz.data()),
-            reinterpret_cast<const char*>(xyz.data() + xyz.size()));
-    pnts.insert(
-            pnts.end(),
-            reinterpret_cast<const char*>(rgb.data()),
-            reinterpret_cast<const char*>(rgb.data() + rgb.size()));
-
-    return pnts;
 }
 
 } // namespace cesium
